+++ purchase_order_line_display_sequence/i18n/de.po
