+++ purchase_order_line_display_sequence/static/description/icon.png
