--- purchase_order_line_sequence/models/purchase_line.py
@@ -1,17 +1,0 @@
-# Copyright 2017 Camptocamp SA - Damien Crier, Alexandre Fayolle
-# Copyright 2017 Eficent Business and IT Consulting Services S.L.
-# Copyright 2017 Serpent Consulting Services Pvt. Ltd.
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
-
-from odoo import models
-
-
-class PurchaseOrderLine(models.Model):
-    _inherit = "purchase.order.line"
-
-    def _prepare_account_move_line(self, move=False):
-        self.ensure_one()
-        res = super(PurchaseOrderLine, self)._prepare_account_move_line(move)
-        res["sequence"] = self.sequence
-
-        return res
