+++ setup/purchase_order_line_display_sequence/odoo/addons/purchase_order_line_display_sequence
@@ -1,0 +1,1 @@
+../../../../purchase_order_line_display_sequence
